MultiDimensional.i: move unsigned long long methods to GUInt64
I looked at the C and C++ api docs and the changed methods seem
to call GDAL routines that return GUInt64's.

It was my hope this would fix the GetSize() bug but it's still
not working.

+++ swig/java/apps/GDALTestMultiDim.java
@@ -30,7 +30,6 @@
  * DEALINGS IN THE SOFTWARE.
  *****************************************************************************/
 
-import java.math.BigInteger;
 import java.util.Arrays;
 
 import org.gdal.gdal.gdal;
@@ -108,7 +107,7 @@ public class GDALTestMultiDim
             
             Dimension d = outDims[i];
 
-            if (d.GetSize().longValue() != sizes[i]) {
+            if (d.GetSize() != sizes[i]) {
 
                 throw new RuntimeException("resulting dimension "+i+" has size "+ d.GetSize()+" but should equal "+sizes[i]);
             }
